requirements.txt
aiodns==3.0.0
aiormq==6.7.7
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgi-lifespan==2.1.0
asyncio==3.4.3
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.3.1
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
execnet==2.0.2
fastapi==0.103.2
filelock==3.12.3
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httptools==0.6.0
hyperframe==6.0.1
identify==2.5.28
idna==3.4
iniconfig==2.0.0
isort==5.12.0
loguru==0.7.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
nodeenv==1.8.0
outcome==1.2.0
packaging==23.1
pamqp==3.2.1
passlib==1.7.4
pathlib==1.0.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
pre-commit==3.4.0
pyasn1==0.5.0
pycares==4.3.0
pycparser==2.21
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.21
starlette==0.27.0
text-unidecode==1.3
trio==0.22.2
typing_extensions==4.7.1
uvicorn==0.23.2
virtualenv==20.24.2
watchfiles==0.20.0
websockets==11.0.3
xmltodict==0.13.0
yarl==1.9.2

envkey~=2.4.2
cryptography~=42.0.5